Kafka_SparkStreaming/workspace/producer/requirements.txt
confluent-kafka==2.3.0
requests==2.31.0
sseclient-py==1.8.0